fix(ios): resolve SwiftLint errors in notebook editor files.
Split oversized views, rename short identifiers, and replace force-try regex init so the iOS CI lint step passes.
