+++ clients/ios/Lextures/Core/Notebook/NotebookDrawing.swift
@@ -4,7 +4,7 @@ import SwiftUI
 /// One whiteboard element from a ```drawing fenced block (parity with web `whiteboard/types`).
 enum NotebookDrawEl: Equatable {
     case stroke(color: String, width: Double, pts: [CGPoint])
-    case rect(color: String, width: Double, x: Double, y: Double, w: Double, h: Double)
+    case rect(color: String, width: Double, rectX: Double, rectY: Double, rectWidth: Double, rectHeight: Double)
     case circle(color: String, width: Double, cx: Double, cy: Double, rx: Double, ry: Double)
     case triangle(color: String, width: Double, x1: Double, y1: Double, x2: Double, y2: Double, x3: Double, y3: Double)
     case line(color: String, width: Double, x1: Double, y1: Double, x2: Double, y2: Double)
@@ -37,16 +37,22 @@ enum NotebookDrawing {
         switch obj["type"] as? String {
         case "stroke":
             let pts = (obj["pts"] as? [[Any]] ?? []).compactMap { pair -> CGPoint? in
-                guard pair.count >= 2, let x = num(pair[0]), let y = num(pair[1]) else { return nil }
-                return CGPoint(x: x, y: y)
+                guard pair.count >= 2, let pointX = num(pair[0]), let pointY = num(pair[1]) else { return nil }
+                return CGPoint(x: pointX, y: pointY)
             }
             return .stroke(color: color, width: width, pts: pts)
         case "rect":
-            guard let x = num(obj["x"]), let y = num(obj["y"]), let w = num(obj["w"]), let h = num(obj["h"]) else { return nil }
-            return .rect(color: color, width: width, x: x, y: y, w: w, h: h)
+            guard
+                let rectX = num(obj["x"]), let rectY = num(obj["y"]),
+                let rectW = num(obj["w"]), let rectH = num(obj["h"])
+            else { return nil }
+            return .rect(color: color, width: width, rectX: rectX, rectY: rectY, rectWidth: rectW, rectHeight: rectH)
         case "circle":
-            guard let cx = num(obj["cx"]), let cy = num(obj["cy"]), let rx = num(obj["rx"]), let ry = num(obj["ry"]) else { return nil }
-            return .circle(color: color, width: width, cx: cx, cy: cy, rx: rx, ry: ry)
+            guard
+                let centerX = num(obj["cx"]), let centerY = num(obj["cy"]),
+                let radiusX = num(obj["rx"]), let radiusY = num(obj["ry"])
+            else { return nil }
+            return .circle(color: color, width: width, cx: centerX, cy: centerY, rx: radiusX, ry: radiusY)
         case "triangle":
             guard
                 let x1 = num(obj["x1"]), let y1 = num(obj["y1"]),
@@ -55,7 +61,10 @@ enum NotebookDrawing {
             else { return nil }
             return .triangle(color: color, width: width, x1: x1, y1: y1, x2: x2, y2: y2, x3: x3, y3: y3)
         case "line":
-            guard let x1 = num(obj["x1"]), let y1 = num(obj["y1"]), let x2 = num(obj["x2"]), let y2 = num(obj["y2"]) else { return nil }
+            guard
+                let x1 = num(obj["x1"]), let y1 = num(obj["y1"]),
+                let x2 = num(obj["x2"]), let y2 = num(obj["y2"])
+            else { return nil }
             return .line(color: color, width: width, x1: x1, y1: y1, x2: x2, y2: y2)
         default:
             return nil
@@ -63,14 +72,14 @@ enum NotebookDrawing {
     }
 
     static func serializeElements(_ elements: [NotebookDrawEl]) -> String {
-        let array: [[String: Any]] = elements.map { el in
-            switch el {
+        let array: [[String: Any]] = elements.map { element in
+            switch element {
             case .stroke(let color, let width, let pts):
                 return ["type": "stroke", "color": color, "width": width, "pts": pts.map { [$0.x, $0.y] }]
-            case .rect(let color, let width, let x, let y, let w, let h):
-                return ["type": "rect", "color": color, "width": width, "x": x, "y": y, "w": w, "h": h]
-            case .circle(let color, let width, let cx, let cy, let rx, let ry):
-                return ["type": "circle", "color": color, "width": width, "cx": cx, "cy": cy, "rx": rx, "ry": ry]
+            case .rect(let color, let width, let rectX, let rectY, let rectWidth, let rectHeight):
+                return ["type": "rect", "color": color, "width": width, "x": rectX, "y": rectY, "w": rectWidth, "h": rectHeight]
+            case .circle(let color, let width, let centerX, let centerY, let radiusX, let radiusY):
+                return ["type": "circle", "color": color, "width": width, "cx": centerX, "cy": centerY, "rx": radiusX, "ry": radiusY]
             case .triangle(let color, let width, let x1, let y1, let x2, let y2, let x3, let y3):
                 return ["type": "triangle", "color": color, "width": width, "x1": x1, "y1": y1, "x2": x2, "y2": y2, "x3": x3, "y3": y3]
             case .line(let color, let width, let x1, let y1, let x2, let y2):
@@ -87,19 +96,19 @@ enum NotebookDrawing {
     static func contentSize(_ elements: [NotebookDrawEl], minSize: CGSize = CGSize(width: 320, height: 220)) -> CGSize {
         var maxX = minSize.width
         var maxY = minSize.height
-        func grow(_ x: Double, _ y: Double) {
-            maxX = max(maxX, x)
-            maxY = max(maxY, y)
+        func grow(_ coordX: Double, _ coordY: Double) {
+            maxX = max(maxX, coordX)
+            maxY = max(maxY, coordY)
         }
-        for el in elements {
-            switch el {
+        for element in elements {
+            switch element {
             case .stroke(_, _, let pts):
-                for p in pts { grow(p.x, p.y) }
-            case .rect(_, _, let x, let y, let w, let h):
-                grow(x + w, y + h)
-                grow(x, y)
-            case .circle(_, _, let cx, let cy, let rx, let ry):
-                grow(cx + rx, cy + ry)
+                for point in pts { grow(point.x, point.y) }
+            case .rect(_, _, let rectX, let rectY, let rectWidth, let rectHeight):
+                grow(rectX + rectWidth, rectY + rectHeight)
+                grow(rectX, rectY)
+            case .circle(_, _, let centerX, let centerY, let radiusX, let radiusY):
+                grow(centerX + radiusX, centerY + radiusY)
             case .triangle(_, _, let x1, let y1, let x2, let y2, let x3, let y3):
                 grow(x1, y1); grow(x2, y2); grow(x3, y3)
             case .line(_, _, let x1, let y1, let x2, let y2):
@@ -123,32 +132,42 @@ enum NotebookDrawing {
 
     /// Draw all elements into a SwiftUI canvas context at the given scale.
     static func draw(_ elements: [NotebookDrawEl], in context: GraphicsContext, scale: CGFloat) {
-        for el in elements {
-            switch el {
+        for element in elements {
+            switch element {
             case .stroke(let color, let width, let pts):
                 guard pts.count > 1 else {
-                    if let p = pts.first {
-                        let r = max(width, 2) * scale / 2
-                        let dot = Path(ellipseIn: CGRect(x: p.x * scale - r, y: p.y * scale - r, width: r * 2, height: r * 2))
+                    if let point = pts.first {
+                        let dotRadius = max(width, 2) * scale / 2
+                        let dot = Path(ellipseIn: CGRect(
+                            x: point.x * scale - dotRadius,
+                            y: point.y * scale - dotRadius,
+                            width: dotRadius * 2,
+                            height: dotRadius * 2
+                        ))
                         context.fill(dot, with: .color(NotebookDrawing.color(from: color)))
                     }
                     continue
                 }
                 var path = Path()
                 path.move(to: CGPoint(x: pts[0].x * scale, y: pts[0].y * scale))
-                for p in pts.dropFirst() {
-                    path.addLine(to: CGPoint(x: p.x * scale, y: p.y * scale))
+                for point in pts.dropFirst() {
+                    path.addLine(to: CGPoint(x: point.x * scale, y: point.y * scale))
                 }
                 context.stroke(
                     path,
                     with: .color(NotebookDrawing.color(from: color)),
                     style: StrokeStyle(lineWidth: width * scale, lineCap: .round, lineJoin: .round)
                 )
-            case .rect(let color, let width, let x, let y, let w, let h):
-                let rect = CGRect(x: x * scale, y: y * scale, width: w * scale, height: h * scale)
+            case .rect(let color, let width, let rectX, let rectY, let rectWidth, let rectHeight):
+                let rect = CGRect(x: rectX * scale, y: rectY * scale, width: rectWidth * scale, height: rectHeight * scale)
                 context.stroke(Path(rect), with: .color(NotebookDrawing.color(from: color)), lineWidth: width * scale)
-            case .circle(let color, let width, let cx, let cy, let rx, let ry):
-                let rect = CGRect(x: (cx - rx) * scale, y: (cy - ry) * scale, width: rx * 2 * scale, height: ry * 2 * scale)
+            case .circle(let color, let width, let centerX, let centerY, let radiusX, let radiusY):
+                let rect = CGRect(
+                    x: (centerX - radiusX) * scale,
+                    y: (centerY - radiusY) * scale,
+                    width: radiusX * 2 * scale,
+                    height: radiusY * 2 * scale
+                )
                 context.stroke(Path(ellipseIn: rect), with: .color(NotebookDrawing.color(from: color)), lineWidth: width * scale)
             case .triangle(let color, let width, let x1, let y1, let x2, let y2, let x3, let y3):
                 var path = Path()
@@ -175,44 +194,47 @@ enum NotebookDrawing {
     }
 
     /// Whether a point (content coordinates) is within `radius` of an element — eraser hit test.
-    static func hitTest(_ el: NotebookDrawEl, point: CGPoint, radius: Double) -> Bool {
-        func distToSegment(_ p: CGPoint, _ a: CGPoint, _ b: CGPoint) -> Double {
-            let dx = b.x - a.x, dy = b.y - a.y
-            let lengthSq = dx * dx + dy * dy
-            if lengthSq == 0 { return Double(hypot(p.x - a.x, p.y - a.y)) }
-            let t = max(0, min(1, ((p.x - a.x) * dx + (p.y - a.y) * dy) / lengthSq))
-            return Double(hypot(p.x - (a.x + t * dx), p.y - (a.y + t * dy)))
+    static func hitTest(_ element: NotebookDrawEl, point: CGPoint, radius: Double) -> Bool {
+        func distToSegment(_ probe: CGPoint, _ start: CGPoint, _ end: CGPoint) -> Double {
+            let deltaX = end.x - start.x, deltaY = end.y - start.y
+            let lengthSq = deltaX * deltaX + deltaY * deltaY
+            if lengthSq == 0 { return Double(hypot(probe.x - start.x, probe.y - start.y)) }
+            let projection = max(0, min(1, ((probe.x - start.x) * deltaX + (probe.y - start.y) * deltaY) / lengthSq))
+            return Double(hypot(
+                probe.x - (start.x + projection * deltaX),
+                probe.y - (start.y + projection * deltaY)
+            ))
         }
-        switch el {
+        switch element {
         case .stroke(_, let width, let pts):
-            let r = radius + width / 2
-            if pts.count == 1 { return Double(hypot(point.x - pts[0].x, point.y - pts[0].y)) <= r }
-            for i in 0 ..< max(0, pts.count - 1) where distToSegment(point, pts[i], pts[i + 1]) <= r {
+            let hitRadius = radius + width / 2
+            if pts.count == 1 { return Double(hypot(point.x - pts[0].x, point.y - pts[0].y)) <= hitRadius }
+            for segIndex in 0 ..< max(0, pts.count - 1) where distToSegment(point, pts[segIndex], pts[segIndex + 1]) <= hitRadius {
                 return true
             }
             return false
         case .line(_, let width, let x1, let y1, let x2, let y2):
             return distToSegment(point, CGPoint(x: x1, y: y1), CGPoint(x: x2, y: y2)) <= radius + width / 2
-        case .rect(_, let width, let x, let y, let w, let h):
+        case .rect(_, let width, let rectX, let rectY, let rectWidth, let rectHeight):
             let corners = [
-                CGPoint(x: x, y: y), CGPoint(x: x + w, y: y),
-                CGPoint(x: x + w, y: y + h), CGPoint(x: x, y: y + h),
+                CGPoint(x: rectX, y: rectY), CGPoint(x: rectX + rectWidth, y: rectY),
+                CGPoint(x: rectX + rectWidth, y: rectY + rectHeight), CGPoint(x: rectX, y: rectY + rectHeight),
             ]
-            let r = radius + width / 2
-            for i in 0 ..< 4 where distToSegment(point, corners[i], corners[(i + 1) % 4]) <= r {
+            let hitRadius = radius + width / 2
+            for cornerIndex in 0 ..< 4 where distToSegment(point, corners[cornerIndex], corners[(cornerIndex + 1) % 4]) <= hitRadius {
                 return true
             }
             return false
-        case .circle(_, let width, let cx, let cy, let rx, let ry):
-            guard rx > 0, ry > 0 else { return false }
-            let nx = (point.x - cx) / rx
-            let ny = (point.y - cy) / ry
-            let norm = Double(hypot(nx, ny))
-            return abs(norm - 1) * min(rx, ry) <= radius + width / 2
+        case .circle(_, let width, let centerX, let centerY, let radiusX, let radiusY):
+            guard radiusX > 0, radiusY > 0 else { return false }
+            let normX = (point.x - centerX) / radiusX
+            let normY = (point.y - centerY) / radiusY
+            let norm = Double(hypot(normX, normY))
+            return abs(norm - 1) * min(radiusX, radiusY) <= radius + width / 2
         case .triangle(_, let width, let x1, let y1, let x2, let y2, let x3, let y3):
             let pts = [CGPoint(x: x1, y: y1), CGPoint(x: x2, y: y2), CGPoint(x: x3, y: y3)]
-            let r = radius + width / 2
-            for i in 0 ..< 3 where distToSegment(point, pts[i], pts[(i + 1) % 3]) <= r {
+            let hitRadius = radius + width / 2
+            for cornerIndex in 0 ..< 3 where distToSegment(point, pts[cornerIndex], pts[(cornerIndex + 1) % 3]) <= hitRadius {
                 return true
             }
             return false
